create ft_strmapi, ft_striteri and adjust same functions

+++ ft_strnstr.c
@@ -6,7 +6,7 @@
 /*   By: tiagalex <tiagalex@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/25 14:32:01 by tiagalex          #+#    #+#             */
-/*   Updated: 2024/10/30 11:47:37 by tiagalex         ###   ########.fr       */
+/*   Updated: 2024/10/31 14:24:12 by tiagalex         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,8 +15,8 @@
 //procura a agulha dentro do palheiro ate um certo limite
 char	*ft_strnstr(const char *haystack, const char *needle, size_t len)
 {
-	int	i;
-	int	j;
+	size_t	i;
+	size_t	j;
 
 	i = 0;
 	j = 0;
